+++ truck-meshalgo/tests/tessellation/triangulation.rs
@@ -18,7 +18,7 @@ fn read_jsons() -> Vec<Vec<u8>> {
 fn solid_is_closed() {
     for (i, json) in read_jsons().into_iter().enumerate() {
         let solid: Solid = serde_json::from_reader(json.as_slice()).unwrap();
-        let mut poly = solid.triangulation(0.02).unwrap().to_polygon();
+        let mut poly = solid.triangulation(0.02).to_polygon();
         poly.put_together_same_attrs()
             .remove_degenerate_faces()
             .remove_unused_attrs();
@@ -35,7 +35,7 @@ fn solid_is_closed() {
 fn compare_occt_mesh() {
     let jsons = read_jsons();
     let solid: Solid = serde_json::from_slice(jsons[2].as_slice()).unwrap();
-    let res = solid.triangulation(0.01).unwrap().to_polygon();
+    let res = solid.triangulation(0.01).to_polygon();
     let path = concat!(dir!(), "../obj/by_occt.obj");
     let ans = obj::read(std::fs::read(path).unwrap().as_slice()).unwrap();
     assert!(res.is_clung_to_by(ans.positions(), 0.05));
@@ -46,5 +46,5 @@ fn compare_occt_mesh() {
 fn large_number_meshing() {
     let json = std::fs::read(concat!(dir!(), "large-torus.json")).unwrap();
     let torus: Solid = serde_json::from_slice(json.as_slice()).unwrap();
-    let _ = torus.triangulation(1.0).unwrap().to_polygon();
+    let _ = torus.triangulation(1.0).to_polygon();
 }