+++ truck-rendimpl/examples/material-samples.rs
@@ -80,7 +80,7 @@ impl App for MyApp {
         let e = builder::tsweep(&v, Vector3::unit_x());
         let f = builder::tsweep(&e, Vector3::unit_y());
         let cube = builder::tsweep(&f, Vector3::unit_z());
-        let mesh = cube.triangulation(0.01).unwrap().to_polygon();
+        let mesh = cube.triangulation(0.01).to_polygon();
         let instance: PolygonInstance = scene
             .instance_creator()
             .create_instance(&mesh, &Default::default());